requirements.txt
argparse
jira
logger
datetime
pyyaml
